Package clean up: reorder files, remove unused code.
