--- src/provider_formatting.jl
@@ -1,17 +1,0 @@
-function process(r::JSONRPC.Request{Val{Symbol("textDocument/formatting")},DocumentFormattingParams}, server)
-    if !haskey(server.documents, URI2(r.params.textDocument.uri))
-        send(JSONRPC.Response(get(r.id), CancelParams(get(r.id))), server)
-        return
-    end
-    doc = server.documents[URI2(r.params.textDocument.uri)]
-    newcontent = DocumentFormat.format(doc._content)
-    end_l, end_c = get_position_at(doc, sizeof(doc._content))
-    lsedits = TextEdit[TextEdit(Range(0, 0, end_l - 1, end_c), newcontent)]
-
-    response = JSONRPC.Response(get(r.id), lsedits)
-    send(response, server)
-end
-
-function JSONRPC.parse_params(::Type{Val{Symbol("textDocument/formatting")}}, params)
-    return DocumentFormattingParams(params)
-end
